+++ e2etests/tests/test_project_uploads.py
@@ -130,14 +130,14 @@ def test_project_upload_external_encodings(requests, a_project, binary_test_file
     assert res.status_code == 201
 
 
-def test_project_upload_external_data(requests, a_blocking_project, binary_test_file_path):
+def test_project_upload_external_data_binary(requests, a_blocking_project, binary_test_file_path):
     project = a_blocking_project
     blocking_data = json.dumps(
-        {str(encoding_id): list({str(encoding_id % 3), str(encoding_id % 13)}) for encoding_id in range(1000)}).encode()
+        {"blocks": {str(encoding_id): list({str(encoding_id % 3), str(encoding_id % 13)}) for encoding_id in range(1000)}}).encode()
 
     mc, upload_info = get_temp_upload_client(project, requests, project['update_tokens'][0])
 
-    _upload_encodings_and_blocks(mc, upload_info, blocking_data, binary_test_file_path)
+    _upload_encodings_and_blocks_binary(mc, upload_info, blocking_data, binary_test_file_path)
 
     # Should be able to notify the service that we've uploaded data
     res = requests.post(url + f"projects/{project['project_id']}/clks",
@@ -182,7 +182,7 @@ def test_project_upload_external_data(requests, a_blocking_project, binary_test_
     assert res2.status_code == 403
 
     mc2, upload_info2 = get_temp_upload_client(project, requests, project['update_tokens'][1])
-    _upload_encodings_and_blocks(mc2, upload_info2, blocking_data, binary_test_file_path)
+    _upload_encodings_and_blocks_binary(mc2, upload_info2, blocking_data, binary_test_file_path)
 
     # If the second data provider uses the correct path to upload data, that should work
     res3 = requests.post(url + f"projects/{project['project_id']}/clks",
@@ -204,19 +204,117 @@ def test_project_upload_external_data(requests, a_blocking_project, binary_test_
                        )
     assert res3.status_code == 201
     run_id = post_run(requests, project, threshold=0.95)
-    result = get_run_result(requests, project, run_id, timeout=120)
+    result = get_run_result(requests, project, run_id, timeout=240)
     assert 'groups' in result
 
 
-def _upload_encodings_and_blocks(mc, upload_info, blocking_data, binary_test_file_path):
+def test_project_upload_external_data_json(requests, a_blocking_project, binary_test_file_path):
+    project = a_blocking_project
+    blocking_data = json.dumps(
+        {"blocks": {str(encoding_id): list({str(encoding_id % 3), str(encoding_id % 13)}) for encoding_id in range(1000)}}).encode()
+
+    clks_data = json.dumps({"clks": ['ssssssss'] * 1000}).encode()
+    mc, upload_info = get_temp_upload_client(project, requests, project['update_tokens'][0])
+
+    _upload_encodings_and_blocks_json(mc, upload_info, blocking_data, clks_data)
+
+    # Should be able to notify the service that we've uploaded data
+    res = requests.post(url + f"projects/{project['project_id']}/clks",
+                        headers={'Authorization': project['update_tokens'][0]},
+                        json={
+                            'encodings': {
+                                'file': {
+                                    'bucket': upload_info['bucket'],
+                                    'path': upload_info['path'] + "/encodings.json",
+                                }
+                            },
+                            'blocks': {
+                                'file': {
+                                    'bucket': upload_info['bucket'],
+                                    'path': upload_info['path'] + "/blocks",
+                                }
+                            }
+
+                        }
+                       )
+    assert res.status_code == 201
+
+    # If the second data provider uses the same path to upload data, that shouldn't work
+    res2 = requests.post(url + f"projects/{project['project_id']}/clks",
+                        headers={'Authorization': project['update_tokens'][1]},
+                        json={
+                            'encodings': {
+                                'file': {
+                                    'bucket': upload_info['bucket'],
+                                    'path': upload_info['path'] + "/encodings.json",
+                                }
+                            },
+                            'blocks': {
+                                'file': {
+                                    'bucket': upload_info['bucket'],
+                                    'path': upload_info['path'] + "/blocks",
+                                }
+                            }
+
+                        }
+                       )
+    assert res2.status_code == 403
+
+    mc2, upload_info2 = get_temp_upload_client(project, requests, project['update_tokens'][1])
+    _upload_encodings_and_blocks_json(mc2, upload_info2, blocking_data, clks_data)
+
+    # If the second data provider uses the correct path to upload data, that should work
+    res3 = requests.post(url + f"projects/{project['project_id']}/clks",
+                        headers={'Authorization': project['update_tokens'][1]},
+                        json={
+                            'encodings': {
+                                'file': {
+                                    'bucket': upload_info2['bucket'],
+                                    'path': upload_info2['path'] + "/encodings.json",
+                                }
+                            },
+                            'blocks': {
+                                'file': {
+                                    'bucket': upload_info2['bucket'],
+                                    'path': upload_info2['path'] + "/blocks",
+                                }
+                            }
+                        }
+                       )
+    assert res3.status_code == 201
+    run_id = post_run(requests, project, threshold=0.95)
+    result = get_run_result(requests, project, run_id, timeout=240)
+    assert 'groups' in result
+
+
+def _upload_encodings_and_blocks_binary(mc, upload_info, blocking_data, binary_test_file_path):
     mc.fput_object(
         upload_info['bucket'],
         upload_info['path'] + "/encodings",
         binary_test_file_path,
         metadata={
             "hash-count": 1000,
-            "hash-size": 128
+            "hash-size": 8
         }
+    )
+    mc.put_object(
+        upload_info['bucket'],
+        upload_info['path'] + "/blocks",
+        io.BytesIO(blocking_data),
+        len(blocking_data)
+    )
+
+
+def _upload_encodings_and_blocks_json(mc, upload_info, blocking_data, clks_data):
+    mc.put_object(
+        upload_info['bucket'],
+        upload_info['path'] + "/encodings.json",
+        io.BytesIO(clks_data),
+        metadata={
+            "hash-count": 1000,
+            "hash-size": 8
+        },
+        length=len(clks_data)
     )
     mc.put_object(
         upload_info['bucket'],
@@ -262,7 +360,7 @@ def test_project_binary_data_uploaded(requests, valid_project_params, binary_tes
             binary_test_file_path, new_project_data['project_id'], token, 1000)
 
     run_id = post_run(requests, new_project_data, 0.99)
-    result = get_run_result(requests, new_project_data, run_id, wait=True)
+    result = get_run_result(requests, new_project_data, run_id, wait=True, timeout=240)
 
     if valid_project_params['result_type'] == 'groups':
         assert 'groups' in result
@@ -307,7 +405,7 @@ def test_project_binary_data_upload_with_different_encoded_size(
             requests, d, project_id, token, 500, size=encoding_size)
 
     run_id = post_run(requests, new_project_data, 0.99)
-    result = get_run_result(requests, new_project_data, run_id, wait=True)
+    result = get_run_result(requests, new_project_data, run_id, wait=True, timeout=240)
     if valid_project_params['result_type'] == 'groups':
         assert 'groups' in result
         groups = result['groups']
@@ -330,7 +428,7 @@ def test_project_json_data_upload_with_various_encoded_sizes(
     )
 
     run_id = post_run(requests, new_project_data, 0.9)
-    result = get_run_result(requests, new_project_data, run_id, wait=True)
+    result = get_run_result(requests, new_project_data, run_id, wait=True, timeout=240)
     if result_type == 'groups':
         assert 'groups' in result
         # This is a pretty bad bound, but we're not testing the
@@ -350,7 +448,7 @@ def test_project_json_data_upload_with_mismatched_encoded_size(
 
     with pytest.raises(AssertionError):
         run_id = post_run(requests, new_project_data, 0.9)
-        get_run_result(requests, new_project_data, run_id, wait=True)
+        get_run_result(requests, new_project_data, run_id, wait=True, timeout=240)
 
 
 def test_project_json_data_upload_with_invalid_encoded_size(
@@ -366,7 +464,7 @@ def test_project_json_data_upload_with_invalid_encoded_size(
 
     with pytest.raises(AssertionError):
         run_id = post_run(requests, new_project_data, 0.9)
-        get_run_result(requests, new_project_data, run_id, wait=True)
+        get_run_result(requests, new_project_data, run_id, wait=True, timeout=240)
 
 
 def test_project_json_data_upload_with_too_small_encoded_size(
@@ -382,7 +480,7 @@ def test_project_json_data_upload_with_too_small_encoded_size(
 
     with pytest.raises(AssertionError):
         run_id = post_run(requests, new_project_data, 0.9)
-        get_run_result(requests, new_project_data, run_id, wait=True)
+        get_run_result(requests, new_project_data, run_id, wait=True, timeout=240)
 
 
 def test_project_json_data_upload_with_too_large_encoded_size(