+++ docs/changelog.rst
@@ -7,6 +7,12 @@ Changelog
 Next Version
 ------------
 
+Version 1.13.0-beta3
+--------------------
+
+- Improved performance for blocks of small size #563
+- fix a problem with the upload to the external object store #564
+- updated documentation #567, $569
 
 Version 1.13.0-beta2
 --------------------
@@ -55,9 +61,6 @@ Version 1.13.0-beta
     - We have enabled `dependabot <https://dependabot.com/>`_ on GitHub to keep our Python dependencies up to date.
     - ``anonlinkclient`` now used for benchmarking. (#490)
     - Chart dependencies ``redis-ha``, ``postgres`` and ``minio`` all updated. (#496, #497)
-
-Breaking Change
-~~~~~~~~~~~~~~~
 
 Breaking Changes
 ~~~~~~~~~~~~~~~~